refactor(outputs): extract CapacityAct (layout unchanged)
Move Capacity Overview, ZFS/Longhorn details, and Backup cards out of
OutputDashboard into a dedicated CapacityAct component, including the
capacitySegments donut-chart data it now owns internally.
